+++ src/index.ts
@@ -4,3 +4,4 @@ export * from "./lib/apply-selector"
 export * from "./lib/types"
 export * from "./lib/utils"
 export * from "./lib/gerber"
+export * from "./lib/excellon-drill"
