+++ src/lib/excellon-drill/commands/M48.ts
@@ -1,0 +1,10 @@
+import { z } from "zod"
+import { defineExcellonDrillCommand } from "../define-excellon-drill-command"
+
+export const M48 = defineExcellonDrillCommand({
+  command_code: "M48",
+  schema: z.object({
+    command_code: z.literal("M48").default("M48"),
+  }),
+  stringify: () => "M48",
+})
