+++ tests/excellon-drill/excellon-drill-builder.test.ts
@@ -1,5 +1,5 @@
 import test from "ava"
-import { excellonDrill } from "lib/excellon-drill"
+import { excellonDrill } from "lib/excellon-drill/excellon-drill-builder"
 
 test("test that we can recreate an example drill file", (t) => {
   const output_commands = excellonDrill()
